requirements.txt
flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0